Allow using multiple keys and selecting between them via kid
JWT contains a mechanism for selecting the correct key by
using an identifier. This commit adds an ability to use a different
key based on the parameter in the JWT to verify.

To make sure that the change doesn't introduce any issues a new test
script was added to run some sanity tests.